refactor(app): merge hooks
